Added ESC configuration
Rev1 was set speeds of 8000, 9000, and 10000 rpm, with conservative
values for spin up + data logging.

Rev2 was set speeds of 3000, 5000, and 7000 rpm, to test low speed
pick up of balls. It was inferior to high speed pickup.

Rev3 was a return to the orginal speed settings, with more daring
spin up and PID tunings.

The .csv files are data logged from the revisions. Persue at your
leisure